Wait for old clients to exit before starting new
During high load it can take a very long time for blockfetch and
chainsync to exit, so long that the governor can re-promote the peer.

By waiting for the previous blockfetch to remove itself from the fetch
registry we know that the old blockfetch and the chainsync client have
exited.
By having chainsync wait for the old chainsync to exit and a signal from
the new blockfetch we know that the two new clients are hooked up.
By having keepalive wait for the old keepalive to exit we know that any
blockfetch and chainsync clients have been killed before we start a new.

Since the startup of blockfetch is split into two steps a new status
called PeerFetchStatusStarting has been added.