Only share PeerGSV info for hot peers
With p2p the keepalive protocol will be running for warm peers too
(peers we currently don't run blockfetch for). It is important that we
don't share the PeerGSV data for these peers because it can propagate
stale information such as request in flight for a failed but now
re-established connection.