Limit the rate with which one can discover all nodes
The commented out function sampleAdvertisedPeers outlined an idea of
limiting the rate with which one can discover all nodes in the network.

This implements that behaviour within computePeerSharingPeers.
Given how cheap it is to create a new connection that can ask for peers
the node will generate one list of peers regardless of who is asking.
The list of peers will change after `policyPeerShareRetryTime` seconds.
The list of peers shared does at most change by the number of peers added or removed.
That is a newly added or removed peer can at most lead to one new peer beeing shared.